moved filtering and tag system into CardGroup
