+++ src/components/CardGroup.module.scss
@@ -6,3 +6,22 @@
   flex-wrap: wrap;
   justify-content: center;
 }
+
+.headerTag {
+  text-align: center;
+}
+
+.headerTag > span {
+  display: inline-block;
+  font-size: 30px;
+  margin: 10px 25px 0 0;
+  color: #ffe75e;
+}
+
+.headerTag > span:hover {
+  cursor: pointer;
+}
+
+.headerTag > span.active {
+  font-weight: bold;
+}
