+++ src/components/FilterProjects.js
@@ -1,10 +1,12 @@
+import styles from "./CardGroup.module.css";
 export default function FilterProjects(props) {
     return (
         <span 
-          className={ props.active ? "active" : "" }
+          className={ props.active ? styles.active : "" } 
           onClick={ () => props.onClickChild(props.tag) }
         >
             {props.tag} 
         </span>
     )
+
 }
