+++ src/index.js
@@ -2,11 +2,8 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.scss';
 import { projects } from './project-data.js';
-import { useState, useEffect } from "react";
 import Nav from './components/Nav'; // import the nav component
-import rocks from "./images/rocks.png";
 import CardGroup from './components/CardGroup';
-import FilterProjects from './components/FilterProjects';
 
 // const cardss = Array.from({ length: 11 }).map(a => 
 //   ({
@@ -22,44 +19,17 @@ import FilterProjects from './components/FilterProjects';
 //     image: rocks
 //   })
 // )
-const tags = ["All", "Misc", "AI", "iOS", "Self Driving", "Computer Vision"];
+
 
 function App() {
-  const [cards, setCards] = useState(null);
-  const [activeIndex, setActiveIndex] = useState(0);
-
-  useEffect(() => {
-    setCards(projects);
-  }, [])
-
-  const filterProjects = (i, filterTag) => {
-    setActiveIndex(i);
-    if (filterTag == 'All') {
-      setCards(projects)
-      return;
-    }
-    const filteredProjects = projects.filter((project) => project.tags.includes(filterTag));
-    setCards(filteredProjects)
-  };
 
   return (
     // shorthand for the React.Fragment
     <>
       <Nav />
       <h1>COMP2022</h1>
-      <div className='headerTag'>
-        { tags.map((tag, index) => 
-            <FilterProjects 
-              tag={ tag } 
-              key={ index }
-              active={ activeIndex === index }
-              onClickChild= { (filterTag) => filterProjects(index, filterTag) }
-            />
-            ) 
-        }
-      </div>
       <div id="container">
-        {cards && <CardGroup cards={cards} />}
+        <CardGroup cards={projects} />
       </div>
     </>
   );
