+++ src/index.scss
@@ -44,7 +44,6 @@ body {
 .headerTag > span:hover {
   cursor: pointer;
 }
-
 .headerTag > span.active {
   font-weight: bold;
 }
